moved typedef declarations to a types header
