+++ include/cudoku.h
@@ -4,13 +4,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include "../include/types.h"
 #include "../include/grids.h"
 #include "../include/constants.h"
 #include "../include/parser.h"
 #include "../include/intersect.h"
 #include "../include/utils.h"
-
-int* getColumn(int **puzzle, int x);
 
 int solve(int **puzzle);
