+++ include/grids.h
@@ -2,12 +2,11 @@
 #define GRIDS_H
 
 #include "../include/constants.h"
+#include "../include/types.h"
+#include "../include/intersect.h"
 #include <stdlib.h>
 
-typedef struct Point {
-    int x;
-    int y;
-} Point;
+int* getColumn(int **puzzle, int x);
 
 int** getGrid(int **puzzle, int x, int y);
 
@@ -15,6 +14,6 @@ Point *getGridCoordinates(int x, int y);
 
 int *getGridArray(int **puzzle, int x, int y);
 
-Point* findObviousPair(int **puzzle, int i, int j);
+Point* findObviousPair(int **puzzle, int i, int j, Set* possibleValues);
 
 #endif
