+++ include/types.h
@@ -1,0 +1,14 @@
+#ifndef TYPES_H
+#define TYPES_H
+
+typedef struct Point {
+    int x;
+    int y;
+} Point;
+
+typedef struct {
+    int *array;
+    int length;
+} Set;
+
+#endif
